+++ R/parsers_basic.R
@@ -399,9 +399,12 @@ eof <- function() {
 #'   p \%then\% zero_or_more(p)
 #'
 #' exactly(n,p):
-#'   counter = 0
-#'   r = zero_or_more((p %using% F(x): counter = counter + 1; x))(x)
-#'   if counter == n then r else fail()(x)
+#'   count = 0
+#'   r = zero_or_more((p %using% F(x): count = count + 1; x))(x)
+#'   if counter == n then
+#'     count = 0
+#'     r
+#'   else fail()(x)
 #'
 #' zero_or_one:
 #'   exactly(0,p) \%or\% exactly(1,p)
@@ -445,10 +448,18 @@ exactly <- function(n, p) {
   # The non-greedy version is match_n
   stopifnot(n >= 0)
   stopifnot(as.integer(n) == n)
-  counter <- 0
+  cnt <- 0
   function(x) {
-    r <- zero_or_more((p) %using% function(x) {counter <<- counter + 1; return(x)})(x)
-    if (counter == n) r else fail()(x)
+    r <- zero_or_more((p) %using% function(x) {cnt <<- cnt + 1; return(x)})(x)
+    if (cnt == n) {
+      # reset cnt
+      cnt <<- 0
+      return(r)
+    }
+    else {
+      # cnt <<- 0
+      return(fail()(x))
+    }
   }
 }